[add]添加顾客信息展示页面（加载失败） 后续还需更新改进

+++ src/main/java/com/lushiqi/kinokocrm/entity/repo/CustomerClassesRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CustomerClassesRepository extends JpaRepository<CustomerClasses, Integer> {
+public interface CustomerClassesRepository extends JpaRepository<CustomerClasses, Long> {
 }
