+++ src/main/java/com/lushiqi/kinokocrm/entity/repo/CustomersRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CustomersRepository extends JpaRepository<Customers, Integer> {
+public interface CustomersRepository extends JpaRepository<Customers, Long> {
 }
